requirements.txt
flask==2.0.1
psycopg2-binary>=2.9.5
python-dotenv==0.19.0
werkzeug==2.1.2
mysql-connector-python==8.0.33 
gunicorn==20.1.0
